Reimplemented camera interface
