+++ .gitignore
@@ -1,3 +1,5 @@
+
+.history
 *.lock
 .env
 # Byte-compiled / optimized / DLL files
