+++ Pipfile
@@ -7,6 +7,8 @@ name = "pypi"
 flask = "*"
 pigpio = "*"
 requests = "*"
+pillow = "*"
+picamera2 = "*"
 
 [dev-packages]
